+++ ogn/commands/logbook.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from datetime import timedelta
+from datetime import timedelta, datetime
 
 from sqlalchemy.sql import func, null
 from sqlalchemy import and_, or_
 from sqlalchemy.sql.expression import true, false, label
+from sqlalchemy.orm import aliased
 
 from ogn.model import Device, TakeoffLanding, Airport
 
@@ -24,8 +25,10 @@ def compute():
     print("New/recalculated takeoffs/landings: {}".format(counter))
 
 
+@manager.arg('date', help='date (format: yyyy-mm-dd')
+@manager.arg('utc_delta_hours', help='delta hours to utc (for local time logs)')
 @manager.command
-def show(airport_name):
+def show(airport_name, utc_delta_hours=0, date=None):
     """Show a logbook for <airport_name>."""
     airport = session.query(Airport) \
         .filter(Airport.name == airport_name) \
@@ -35,110 +38,152 @@ def show(airport_name):
         print('Airport "{}" not found.'.format(airport_name))
         return
 
-    delta_altitude = 200
-    delta_radius = 0.01  # degree!
+    utc_timedelta = timedelta(hours=utc_delta_hours)
+    or_args = []
+    if date is not None:
+        date = datetime.strptime(date, "%Y-%m-%d")
+        or_args = [and_(TakeoffLanding.timestamp >= date + utc_timedelta,
+                        TakeoffLanding.timestamp < date + timedelta(hours=24) + utc_timedelta)]
 
     # make a query with current, previous and next "takeoff_landing" event, so we can find complete flights
     sq = session.query(
-        TakeoffLanding.address,
-        func.lag(TakeoffLanding.address)
-            .over(
-                order_by=and_(func.date(TakeoffLanding.timestamp),
-                              TakeoffLanding.address,
-                              TakeoffLanding.timestamp))
-            .label('address_prev'),
-        func.lead(TakeoffLanding.address)
-            .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                TakeoffLanding.address,
-                                TakeoffLanding.timestamp))
-            .label('address_next'),
-        TakeoffLanding.timestamp,
+        TakeoffLanding.device_id,
+        func.lag(TakeoffLanding.device_id)
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('device_id_prev'),
+        func.lead(TakeoffLanding.device_id)
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('device_id_next'),
+        (TakeoffLanding.timestamp + utc_timedelta).label('timestamp'),
         func.lag(TakeoffLanding.timestamp)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('timestamp_prev'),
-        func.lead(TakeoffLanding.timestamp)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('timestamp_next'),
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('timestamp_prev'),
+        func.lead(TakeoffLanding.timestamp + utc_timedelta)
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('timestamp_next'),
         TakeoffLanding.track,
         func.lag(TakeoffLanding.track)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('track_prev'),
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('track_prev'),
         func.lead(TakeoffLanding.track)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('track_next'),
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('track_next'),
         TakeoffLanding.is_takeoff,
         func.lag(TakeoffLanding.is_takeoff)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('is_takeoff_prev'),
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('is_takeoff_prev'),
         func.lead(TakeoffLanding.is_takeoff)
-                .over(order_by=and_(func.date(TakeoffLanding.timestamp),
-                                    TakeoffLanding.address,
-                                    TakeoffLanding.timestamp))
-                .label('is_takeoff_next')) \
-        .filter(func.ST_DFullyWithin(TakeoffLanding.location_wkt, Airport.location_wkt, delta_radius)) \
-        .filter(TakeoffLanding.altitude < Airport.altitude + delta_altitude) \
-        .filter(Airport.name == airport.name) \
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('is_takeoff_next'),
+        TakeoffLanding.airport_id,
+        func.lag(TakeoffLanding.airport_id)
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('airport_id_prev'),
+        func.lead(TakeoffLanding.airport_id)
+            .over(order_by=and_(func.date(TakeoffLanding.timestamp + utc_timedelta),
+                                TakeoffLanding.device_id,
+                                TakeoffLanding.timestamp + utc_timedelta))
+            .label('airport_id_next')) \
+        .filter(*or_args) \
         .subquery()
 
-    # find complete flights (with takeoff and landing) with duration < 1 day
-    complete_flight_query = session.query(sq.c.timestamp.label('reftime'), sq.c.address.label('address'), sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), sq.c.timestamp_next.label('landing'), sq.c.track_next.label('landing_track'), label('duration', sq.c.timestamp_next - sq.c.timestamp)) \
+    # find complete flights (with takeoff and landing on the same day)
+    complete_flight_query = session.query(sq.c.timestamp.label('reftime'),
+                                          sq.c.device_id.label('device_id'),
+                                          sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), sq.c.airport_id.label('takeoff_airport_id'),
+                                          sq.c.timestamp_next.label('landing'), sq.c.track_next.label('landing_track'), sq.c.airport_id_next.label('landing_airport_id'),
+                                          label('duration', sq.c.timestamp_next - sq.c.timestamp)) \
         .filter(and_(sq.c.is_takeoff == true(), sq.c.is_takeoff_next == false())) \
-        .filter(sq.c.address == sq.c.address_next) \
-        .filter(sq.c.timestamp_next - sq.c.timestamp < timedelta(days=1))
+        .filter(sq.c.device_id == sq.c.device_id_next) \
+        .filter(func.date(sq.c.timestamp_next) == func.date(sq.c.timestamp)) \
+        .filter(or_(sq.c.airport_id == airport.id,
+                    sq.c.airport_id_next == airport.id))
 
-    # split complete flights (with takeoff and landing) with duration > 1 day into one takeoff and one landing
-    split_start_query = session.query(sq.c.timestamp.label('reftime'), sq.c.address.label('address'), sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), null().label('landing'), null().label('landing_track'), null().label('duration')) \
+    # split complete flights (with takeoff and landing on different days) into one takeoff and one landing
+    split_start_query = session.query(sq.c.timestamp.label('reftime'),
+                                      sq.c.device_id.label('device_id'),
+                                      sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), sq.c.airport_id.label('takeoff_airport_id'),
+                                      null().label('landing'), null().label('landing_track'), null().label('landing_airport_id'),
+                                      null().label('duration')) \
         .filter(and_(sq.c.is_takeoff == true(), sq.c.is_takeoff_next == false())) \
-        .filter(sq.c.address == sq.c.address_next) \
-        .filter(sq.c.timestamp_next - sq.c.timestamp >= timedelta(days=1))
+        .filter(sq.c.device_id == sq.c.device_id_next) \
+        .filter(func.date(sq.c.timestamp_next) != func.date(sq.c.timestamp)) \
+        .filter(and_(sq.c.airport_id == airport.id,
+                     sq.c.airport_id_next == airport.id))
 
-    split_landing_query = session.query(sq.c.timestamp_next.label('reftime'), sq.c.address.label('address'), null().label('takeoff'), null().label('takeoff_track'), sq.c.timestamp_next.label('landing'), sq.c.track_next.label('landing_track'), null().label('duration')) \
+    split_landing_query = session.query(sq.c.timestamp_next.label('reftime'),
+                                        sq.c.device_id.label('device_id'),
+                                        null().label('takeoff'), null().label('takeoff_track'), null().label('takeoff_airport_id'),
+                                        sq.c.timestamp_next.label('landing'), sq.c.track_next.label('landing_track'), sq.c.airport_id_next.label('landing_airport_id'),
+                                        null().label('duration')) \
         .filter(and_(sq.c.is_takeoff == true(), sq.c.is_takeoff_next == false())) \
-        .filter(sq.c.address == sq.c.address_next) \
-        .filter(sq.c.timestamp_next - sq.c.timestamp >= timedelta(days=1))
+        .filter(sq.c.device_id == sq.c.device_id_next) \
+        .filter(func.date(sq.c.timestamp_next) != func.date(sq.c.timestamp)) \
+        .filter(and_(sq.c.airport_id == airport.id,
+                     sq.c.airport_id_next == airport.id))
 
     # find landings without start
-    only_landings_query = session.query(sq.c.timestamp.label('reftime'), sq.c.address.label('address'), null().label('takeoff'), null().label('takeoff_track'), sq.c.timestamp.label('landing'), sq.c.track_next.label('landing_track'), null().label('duration')) \
+    only_landings_query = session.query(sq.c.timestamp.label('reftime'),
+                                        sq.c.device_id.label('device_id'),
+                                        null().label('takeoff'), null().label('takeoff_track'), null().label('takeoff_airport_id'),
+                                        sq.c.timestamp.label('landing'), sq.c.track_next.label('landing_track'), sq.c.airport_id_next.label('landing_airport_id'),
+                                        null().label('duration')) \
         .filter(sq.c.is_takeoff == false()) \
-        .filter(or_(sq.c.address != sq.c.address_prev,
-                    sq.c.is_takeoff_prev == false()))
+        .filter(or_(sq.c.device_id != sq.c.device_id_prev,
+                    sq.c.is_takeoff_prev == false())) \
+        .filter(sq.c.airport_id_next == airport.id)
 
     # find starts without landing
-    only_starts_query = session.query(sq.c.timestamp.label('reftime'), sq.c.address.label('address'), sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), null().label('landing'), null().label('landing_track'), null().label('duration')) \
+    only_starts_query = session.query(sq.c.timestamp.label('reftime'),
+                                      sq.c.device_id.label('device_id'),
+                                      sq.c.timestamp.label('takeoff'), sq.c.track.label('takeoff_track'), sq.c.airport_id.label('takeoff_airport_id'),
+                                      null().label('landing'), null().label('landing_track'), null().label('landing_airport_id'),
+                                      null().label('duration')) \
         .filter(sq.c.is_takeoff == true()) \
-        .filter(or_(sq.c.address != sq.c.address_next,
-                    sq.c.is_takeoff_next == true()))
+        .filter(or_(sq.c.device_id != sq.c.device_id_next,
+                    sq.c.is_takeoff_next == true())) \
+        .filter(sq.c.airport_id == airport.id)
 
     # unite all
-    union_query = complete_flight_query.union(
-        split_start_query,
-        split_landing_query,
-        only_landings_query,
-        only_starts_query) \
+    union_query = complete_flight_query.union(split_start_query,
+                                              split_landing_query,
+                                              only_landings_query,
+                                              only_starts_query) \
         .subquery()
 
-    # get aircraft informations and sort all entries by the reference time
-    logbook_query = session.query(
-        union_query.c.reftime,
-        union_query.c.address,
-        union_query.c.takeoff,
-        union_query.c.takeoff_track,
-        union_query.c.landing,
-        union_query.c.landing_track,
-        union_query.c.duration,
-        Device.registration,
-        Device.aircraft) \
-        .outerjoin(Device, union_query.c.address == Device.address) \
+    # get aircraft and airport informations and sort all entries by the reference time
+    takeoff_airport = aliased(Airport, name='takeoff_airport')
+    landing_airport = aliased(Airport, name='landing_airport')
+    logbook_query = session.query(union_query.c.reftime,
+                                  union_query.c.takeoff,
+                                  union_query.c.takeoff_track,
+                                  takeoff_airport,
+                                  union_query.c.landing,
+                                  union_query.c.landing_track,
+                                  landing_airport,
+                                  union_query.c.duration,
+                                  Device) \
+        .outerjoin(Device, union_query.c.device_id == Device.id) \
+        .outerjoin(takeoff_airport, union_query.c.takeoff_airport_id == takeoff_airport.id) \
+        .outerjoin(landing_airport, union_query.c.landing_airport_id == landing_airport.id) \
         .order_by(union_query.c.reftime)
 
     print('--- Logbook ({}) ---'.format(airport_name))
@@ -152,19 +197,28 @@ def show(airport_name):
     def none_timedelta_replacer(timedelta_object):
         return '--:--:--' if timedelta_object is None else timedelta_object
 
-    def none_registration_replacer(registration_object, address):
-        return '[' + address + ']' if registration_object is None else registration_object
+    def none_registration_replacer(device_object):
+        return '[' + device_object.address + ']' if device_object.registration is None else device_object.registration
 
-    def none_aircraft_replacer(aircraft_object):
-        return '(unknown)' if aircraft_object is None else aircraft_object
+    def none_aircraft_replacer(device_object):
+        return '(unknown)' if device_object.aircraft is None else device_object.aircraft
 
-    for [reftime, address, takeoff, takeoff_track, landing, landing_track, duration, registration, aircraft] in logbook_query.all():
-        print('%10s   %8s (%2s)   %8s (%2s)   %8s   %8s   %s' % (
+    def airport_marker(takeoff_airport_object, landing_airport_object):
+        if takeoff_airport_object is not None and takeoff_airport_object.name is not airport.name:
+            return ('FROM: {}'.format(takeoff_airport_object.name))
+        elif landing_airport_object is not None and landing_airport_object.name is not airport.name:
+            return ('TO: {}'.format(landing_airport_object.name))
+        else:
+            return ('')
+
+    for [reftime, takeoff, takeoff_track, takeoff_airport, landing, landing_track, landing_airport, duration, device] in logbook_query.all():
+        print('%10s   %8s (%2s)   %8s (%2s)   %8s   %8s   %17s %20s' % (
             reftime.date(),
             none_datetime_replacer(takeoff),
             none_track_replacer(takeoff_track),
             none_datetime_replacer(landing),
             none_track_replacer(landing_track),
             none_timedelta_replacer(duration),
-            none_registration_replacer(registration, address),
-            none_aircraft_replacer(aircraft)))
+            none_registration_replacer(device),
+            none_aircraft_replacer(device),
+            airport_marker(takeoff_airport, landing_airport)))