+++ ogn/model/address_origin.py
@@ -2,9 +2,10 @@ class AddressOrigin:
     ogn_ddb = 1
     flarmnet = 2
     user_defined = 3
+    seen = 4
 
     def __init__(self, origin):
-        if origin in [1, 2, 3]:
+        if origin in [1, 2, 3, 4]:
             self.origin = origin
         else:
             raise ValueError('no address origin with id {} known'.format(origin))
@@ -16,4 +17,6 @@ class AddressOrigin:
             return 'FlarmNet'
         elif self.origin == self.user_defined:
             return 'user-defined'
+        elif self.origin == self.seen:
+            return 'seen'
         return ''