+++ ogn/model/aircraft_beacon.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column, String, Integer, Float, Boolean, SmallInteger
+from sqlalchemy import Column, String, Integer, Float, Boolean, SmallInteger, ForeignKey
+from sqlalchemy.orm import relationship
 
 from .beacon import Beacon
 
@@ -8,9 +9,9 @@ class AircraftBeacon(Beacon):
 
     # Flarm specific data
     address_type = Column(SmallInteger)
-    aircraft_type = Column(SmallInteger)
+    aircraft_type = Column(SmallInteger, index=True)
     stealth = Column(Boolean)
-    address = Column(String(6), index=True)
+    address = Column(String(6))
     climb_rate = Column(Float)
     turn_rate = Column(Float)
     signal_strength = Column(Float)
@@ -23,6 +24,13 @@ class AircraftBeacon(Beacon):
     real_address = Column(String(6))
 
     flightlevel = Column(Float)
+
+    # Relations
+    receiver_id = Column(Integer, ForeignKey('receiver.id', ondelete='SET NULL'), index=True)
+    receiver = relationship('Receiver', foreign_keys=[receiver_id])
+
+    device_id = Column(Integer, ForeignKey('device.id', ondelete='SET NULL'), index=True)
+    device = relationship('Device', foreign_keys=[device_id])
 
     def __repr__(self):
         return "<AircraftBeacon %s: %s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s>" % (