+++ ogn/model/airport.py
@@ -1,4 +1,5 @@
 from sqlalchemy import Column, String, Integer, Float, SmallInteger
+from sqlalchemy.orm import relationship
 from geoalchemy2.types import Geometry
 
 from .base import Base
@@ -13,13 +14,16 @@ class Airport(Base):
     altitude = Column(Integer)
 
     name = Column(String, index=True)
-    code = Column(String(5))
+    code = Column(String(6))
     country_code = Column(String(2))
     style = Column(SmallInteger)
     description = Column(String)
     runway_direction = Column(Integer)
     runway_length = Column(Integer)
     frequency = Column(Float)
+
+    # Relations
+    takeoff_landings = relationship('TakeoffLanding')
 
     def __repr__(self):
         return "<Airport %s: %s,%s,%s,%s,%s,%s,%s,%s,%s,% s>" % (